+++ driver-core/src/main/java/com/datastax/driver/core/querybuilder/Utils.java
@@ -23,6 +23,8 @@ import java.util.*;
 import java.util.regex.Pattern;
 
 import com.datastax.driver.core.DataType;
+import com.datastax.driver.core.TupleValue;
+import com.datastax.driver.core.UDTValue;
 import com.datastax.driver.core.utils.Bytes;
 
 // Static utilities private to the query builder
@@ -99,6 +101,12 @@ abstract class Utils {
             return sb;
 
         if (appendValueIfCollection(value, sb))
+            return sb;
+
+        if (appendValueIfUdt(value, sb))
+            return sb;
+
+        if (appendValueIfTuple(value, sb))
             return sb;
 
         appendStringIfValid(value, sb);
@@ -227,6 +235,24 @@ abstract class Utils {
         }
         sb.append('}');
         return sb;
+    }
+
+    private static boolean appendValueIfUdt(Object value, StringBuilder sb) {
+        if (value instanceof UDTValue) {
+            sb.append(((UDTValue)value).toString());
+            return true;
+        } else {
+            return false;
+        }
+    }
+
+    private static boolean appendValueIfTuple(Object value, StringBuilder sb) {
+        if (value instanceof TupleValue) {
+            sb.append(((TupleValue)value).toString());
+            return true;
+        } else {
+            return false;
+        }
     }
 
     static boolean containsBindMarker(Object value) {