+++ driver-core/src/test/java/com/datastax/driver/core/CCMBridge.java
@@ -34,6 +34,8 @@ import java.net.InetSocketAddress;
 import java.util.Collection;
 import java.util.Map;
 
+import static com.google.common.base.Preconditions.checkArgument;
+
 import static com.datastax.driver.core.TestUtils.CREATE_KEYSPACE_SIMPLE_FORMAT;
 import static com.datastax.driver.core.TestUtils.SIMPLE_KEYSPACE;
 
@@ -73,18 +75,25 @@ public class CCMBridge {
     }
 
     public static CCMBridge create(String name) {
+        // This leads to a confusing CCM error message so check explicitly:
+        checkArgument(!"current".equals(name.toLowerCase()),
+                      "cluster can't be called \"current\"");
         CCMBridge bridge = new CCMBridge();
         bridge.execute("ccm create %s -b -i %s %s", name, IP_PREFIX, CASSANDRA_VERSION);
         return bridge;
     }
 
     public static CCMBridge create(String name, int nbNodes) {
+        checkArgument(!"current".equals(name.toLowerCase()),
+                        "cluster can't be called \"current\"");
         CCMBridge bridge = new CCMBridge();
         bridge.execute("ccm create %s -n %d -s -i %s -b %s", name, nbNodes, IP_PREFIX, CASSANDRA_VERSION);
         return bridge;
     }
 
     public static CCMBridge create(String name, int nbNodesDC1, int nbNodesDC2) {
+        checkArgument(!"current".equals(name.toLowerCase()),
+                        "cluster can't be called \"current\"");
         CCMBridge bridge = new CCMBridge();
         bridge.execute("ccm create %s -n %d:%d -s -i %s -b %s", name, nbNodesDC1, nbNodesDC2, IP_PREFIX, CASSANDRA_VERSION);
         return bridge;
@@ -328,6 +337,10 @@ public class CCMBridge {
                 throw new IllegalArgumentException();
 
             return new CCMCluster(CCMBridge.create("test", nbNodesDC1, nbNodesDC2), builder, nbNodesDC1 + nbNodesDC2);
+        }
+
+        public static CCMCluster create(CCMBridge cassandraCluster, Cluster.Builder builder, int totalNodes) {
+            return new CCMCluster(cassandraCluster, builder, totalNodes);
         }
 
         private CCMCluster(CCMBridge cassandraCluster, Cluster.Builder builder, int totalNodes) {