+++ driver-core/CHANGELOG.rst
@@ -5,6 +5,8 @@ CHANGELOG
 ------
 
 - [bug] Release connections on ResultSetFuture#cancel (JAVA-407)
+- [bug] Fix handling of SimpleStatement with values in query builder
+  batches (JAVA-393)
 
 
 2.0.4: