+++ driver-core/CHANGELOG.rst
@@ -8,6 +8,7 @@ CHANGELOG
 - [bug] Fix handling of SimpleStatement with values in query builder
   batches (JAVA-393)
 - [bug] Ensure pool is properly closed in onDown (JAVA-417)
+- [bug] Fix tokenMap initialization at startup (JAVA-415)
 
 
 2.0.4: