+++ driver-core/CHANGELOG.rst
@@ -9,6 +9,7 @@ CHANGELOG
   batches (JAVA-393)
 - [bug] Ensure pool is properly closed in onDown (JAVA-417)
 - [bug] Fix tokenMap initialization at startup (JAVA-415)
+- [bug] Avoid deadlock on close (JAVA-418)
 
 
 2.0.4: