+++ driver-core/CHANGELOG.rst
@@ -15,6 +15,7 @@ Merged from 2.0 branch:
   batches (JAVA-393)
 - [bug] Ensure pool is properly closed in onDown (JAVA-417)
 - [bug] Fix tokenMap initialization at startup (JAVA-415)
+- [bug] Avoid deadlock on close (JAVA-418)
 
 
 2.1.0-rc1: