+++ driver-core/CHANGELOG.rst
@@ -73,6 +73,8 @@ Merged from 2.0 branch: everything up to 2.0.3 (included), and the following.
   (JAVA-368)
 - [bug] Signal connection failure sooner to avoid missing them
   (JAVA-367)
+- [bug] Throw UnsupportedOperationException for protocol batch
+  setSerialCL (JAVA-377)
 
 Merged from 1.0 branch: